requirement.txt
django-htmx-1.12.2
pint